+++ Assets/Scenes/MenuScene.unity
@@ -568,8 +568,7 @@ MonoBehaviour:
       m_Calls: []
   m_text: "\u041F\u0440\u0438\u043B\u043E\u0436\u0435\u043D\u0438\u0435 \u0441\u043E\u0437\u0434\u0430\u043D\u043E\n\u0411\u0443\u043B\u0430\u0435\u0432\u044B\u043C
     \u0410\u043B\u0435\u043A\u0441\u0435\u0435\u043C\n\u043F\u0440\u043E\u0445\u043E\u0434\u044F\u0449\u0438\u0439
-    \u043A\u0443\u0440\u0441\u0441 \u0441\u0442\u0430\u0436\u0438\u0440\u043E\u0432\u043A\u0438!!!\n\n
-    -=IJunior=-"
+    \u043A\u0443\u0440\u0441\u0441 \u0441\u0442\u0430\u0436\u0438\u0440\u043E\u0432\u043A\u0438!!!\n\n-=IJunior=-"
   m_isRightToLeft: 0
   m_fontAsset: {fileID: 11400000, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
   m_sharedMaterial: {fileID: 2180264, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
@@ -638,9 +637,9 @@ MonoBehaviour:
   m_margin: {x: 0, y: -1.4302856, z: 0, w: 0}
   m_textInfo:
     textComponent: {fileID: 932381337}
-    characterCount: 81
+    characterCount: 80
     spriteCount: 0
-    spaceCount: 9
+    spaceCount: 8
     wordCount: 10
     linkCount: 0
     lineCount: 6
@@ -681,7 +680,6 @@ GameObject:
   - component: {fileID: 1022346968}
   - component: {fileID: 1022346967}
   - component: {fileID: 1022346966}
-  - component: {fileID: 1022346965}
   m_Layer: 5
   m_Name: ButtonExit
   m_TagString: Untagged
@@ -709,18 +707,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0.000061035, y: -242}
   m_SizeDelta: {x: 576.08777, y: 109.15289}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &1022346965
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1022346963}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 365af718ea01bc7459753e9e9f2172cb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &1022346966
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -764,7 +750,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1022346965}
+      - m_Target: {fileID: 1702236155}
         m_MethodName: Exit
         m_Mode: 1
         m_Arguments:
@@ -1199,6 +1185,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1702236154}
+  - component: {fileID: 1702236155}
   m_Layer: 5
   m_Name: Menu
   m_TagString: Untagged
@@ -1228,6 +1215,18 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &1702236155
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1702236153}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7a720d2c27f7d5d408b74c83137a8eac, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1973829858
 GameObject:
   m_ObjectHideFlags: 0
@@ -1599,7 +1598,6 @@ GameObject:
   - component: {fileID: 2107650075}
   - component: {fileID: 2107650074}
   - component: {fileID: 2107650073}
-  - component: {fileID: 2107650072}
   m_Layer: 5
   m_Name: ButtonPlay
   m_TagString: Untagged
@@ -1627,18 +1625,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 240}
   m_SizeDelta: {x: 576.08777, y: 109.15289}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &2107650072
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2107650070}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: be11589a5f9048f4384c37e8cc9233d5, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &2107650073
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1682,7 +1668,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 2107650072}
+      - m_Target: {fileID: 1702236155}
         m_MethodName: LoadGame
         m_Mode: 1
         m_Arguments:
