--- Assets/Scripts/UI/ExitGame.cs
@@ -1,10 +1,0 @@
-﻿
-using UnityEngine;
-
-public class ExitGame : MonoBehaviour
-{
-    public void Exit()
-    {
-        Application.Quit();
-    }
-}
