--- Assets/Scripts/UI/LoadScenGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: be11589a5f9048f4384c37e8cc9233d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
