--- Assets/Scripts/UI/LoadScenMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 97fdb4f42c727aa4cbef2eacdeedade2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
